% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cat_variable.R
\name{check_cat_variable}
\alias{check_cat_variable}
\title{Check Categorical Variable Frequencies and Plot}
\usage{
check_cat_variable(data, var, show_plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the variable.}

\item{var}{An unquoted column name of a character or factor variable.}

\item{show_plot}{Logical. Whether to show a bar plot. Default is \code{TRUE}.}
}
\value{
Invisibly returns \code{NULL} after printing output and plotting (if applicable).
}
\description{
Prints a frequency table and missing value count for a categorical variable (character or factor),
and optionally shows a bar plot of the top 10 categories. If more than 10 categories exist,
less frequent ones are grouped into "Other".
}
\examples{
check_cat_variable(Data, f_coffee_farm_age, show_plot = FALSE)

}
