% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_variable.R
\name{check_num_variable}
\alias{check_num_variable}
\title{Check Numeric Variable Summary and Distribution}
\usage{
check_num_variable(data, var, show_plot = TRUE, bins = NULL)
}
\arguments{
\item{data}{A data frame containing the variable.}

\item{var}{An unquoted column name of a numeric variable.}

\item{show_plot}{Logical. Whether to show a histogram. Default is \code{TRUE}.}

\item{bins}{Optional. Number of bins to use in the histogram. If \code{NULL}, it is calculated automatically.}
}
\value{
Invisibly returns \code{NULL} after printing output and plotting (if applicable).
}
\description{
Prints summary statistics, missing value info, and optionally a histogram for a numeric variable.
Automatically calculates histogram bins using the Freedman–Diaconis rule unless specified manually.
}
\examples{
check_num_variable(Data, f_focus_quant_prod, bins = 50)

}
