% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_quantities.R
\name{correct_quantities}
\alias{correct_quantities}
\title{Correct Zero Quantities for Production, Sales, Loss, and optionally Premium Sales}
\usage{
correct_quantities(data, include_premium = TRUE)
}
\arguments{
\item{data}{A data frame with columns: \code{f_focus_quant_prod}, \code{f_focus_quant_sold},
\code{f_focus_quant_lost}, and optionally \code{f_focus_quant_sold_premium}.}

\item{include_premium}{Logical. Whether to include \code{f_focus_quant_sold_premium} in the correction. Default is TRUE.}
}
\value{
A modified data frame with corrections applied and messages summarizing changes.
}
\description{
Replaces 0 values in quantity-related columns with NA when both production and sales are 0.
}
\examples{
Data <- correct_quantities(Data, include_premium = FALSE)

}
