% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outlier_detection}
\alias{outlier_detection}
\title{Outlier Detection Across All Numeric Non-Binary Variables (with Summary)}
\usage{
outlier_detection(data, replacement = 9997, exclude_columns = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{replacement}{The value to replace outliers with. Default is 9997.}

\item{exclude_columns}{Optional character vector of column names to exclude from outlier detection.}
}
\value{
A modified data frame with outliers replaced.
}
\description{
Detects outliers in all numeric, non-binary columns using the mean ± 3*SD rule
and replaces them with a specified value. Prints a summary of how many values were replaced.
}
\examples{
Data <- outlier_detection(Data, replacement = 9997, exclude_columns = c("mobile_number_farmer", "pi_geolocation"))

}
